requirements.txt
GitPython==3.0.5
packaging==19.2
